Changes to user and article routes / controllers

+++ controllers/article.js
@@ -1,21 +1,21 @@
 const Article = require("../models/article");
 const { validationResult, body } = require("express-validator");
 
-exports.GET_Article = (req, res) => {
+exports.GET_Article = (req, res, next) => {
   // Add logic to account for whether article is published or not
   Article.findById(req.params.id)
     .populate("author")
-    .exec(err, (article) => {
+    .exec((err, article) => {
       if (err) return next(err);
       res.status(200).json({ article });
     });
 };
 
-exports.GET_Articles = (req, res) => {
+exports.GET_Articles = (req, res, next) => {
   // Add logic to account for whether articles are published or not
   Article.find()
     .populate("author")
-    .exec(err, (articles) => {
+    .exec((err, articles) => {
       if (err) return next(err);
       res.status(200).json({ articles });
     });
@@ -33,7 +33,14 @@ exports.POST_Article = [
     .escape()
     .withMessage("Text must be at least 20 characters"),
 
-  (req, res) => {
+  (req, res, next) => {
+    if (!req.user.admin) {
+      return res.status(401).json({
+        success: true,
+        message: "You are not authorised to perform this action.",
+      });
+    }
+
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
       return res.status(400).json({ errors: errors.array() });
@@ -45,7 +52,7 @@ exports.POST_Article = [
       author: req.user._id,
     });
 
-    article.save(err, (article) => {
+    article.save().then((err, article) => {
       if (err) return next(err);
       res.status(201).json(article);
     });
@@ -65,6 +72,12 @@ exports.PATCH_Article = [
     .withMessage("Text must be at least 20 characters"),
 
   (req, res) => {
+    if (!req.user.admin) {
+      return res.status(401).json({
+        success: true,
+        message: "You are not authorised to perform this action.",
+      });
+    }
     const errors = validationResult(req);
 
     const updates = {
@@ -75,23 +88,29 @@ exports.PATCH_Article = [
     if (!errors.isEmpty()) {
       res.status(400).json({ message: errors });
     } else {
-      Article.findByIdAndUpdate(req.params.id, article)
+      Article.findByIdAndUpdate(req.params.id, updates)
         .then(() => {
-          res.status(200).json({ message: article });
+          res.status(200).json({ success: true, message: updates });
         })
         .catch((err) => {
-          res.status(400).json({ message: errors });
+          res.status(400).json({ success: false, message: errors });
         });
     }
   },
 ];
 
 exports.DELETE_Article = (req, res) => {
+  if (!req.user.admin) {
+    return res.status(401).json({
+      success: true,
+      message: "You are not authorised to perform this action.",
+    });
+  }
   Article.findByIdAndDelete(req.params.id, (err, article) => {
     if (err) {
-      res.status(400).json({ message: errors });
+      res.status(400).json({ success: false, message: errors });
     } else {
-      res.status(200).json({ message: "Article deleted" });
+      res.status(200).json({ success: true, message: "Article deleted" });
     }
   });
 };
